Adding tests
